+++ src/briefy/ws/utils/user.py
@@ -1,9 +1,11 @@
 """User utilities for briefy.webservice."""
 from briefy.common.utils.cache import timeout_cache
+from briefy.ws import logger
 from briefy.ws.config import USER_SERVICE_BASE
 from briefy.ws.config import USER_SERVICE_TIMEOUT
 
 import requests
+import transaction
 
 
 def _get_user_info_from_service(user_id: str) -> dict:
@@ -19,13 +21,22 @@ def _get_user_info_from_service(user_id: str) -> dict:
     )
     # TODO: improve this to user current user locale
     headers = {'X-Locale': 'en_GB'}
-    resp = requests.get(
-        endpoint,
-        headers=headers
-    )
-    if resp.status_code == 200:
-        raw_data = resp.json()
-        data = raw_data['data'] if 'data' in raw_data else data
+    savepoint = transaction.savepoint()
+    try:
+        resp = requests.get(
+            endpoint,
+            headers=headers
+        )
+    except ConnectionError as exc:
+        logger.warn('Failure connecting to internal user service. Exception: {exc}'.format(exc=exc))
+        savepoint.rollback()
+    else:
+        if resp.status_code == 200:
+            raw_data = resp.json()
+            data = raw_data['data'] if 'data' in raw_data else data
+        else:
+            msg = 'Getting user info from internal services fail. Status code: {status_code}.'
+            logger.info(msg.format(status_code=resp.status_code))
     return data
 
 
@@ -52,13 +63,14 @@ def get_public_user_info(user_id: str) -> dict:
 
 
 def add_user_info_to_state_history(state_history):
-    """Receives object state history and add user information.
+    """Receive object state history and add user information.
 
     :param state_history: list of workflow state history.
     """
     for item in state_history:
-        user_id = item.get('actor', None)
-        if user_id:
+        user = item.get('actor', None)
+        if isinstance(user, str):
+            user_id = user  # first call where actor is a UUID string
             new_actor = get_public_user_info(user_id)
             if new_actor:
                 item['actor'] = new_actor