+++ HISTORY.rst
@@ -2,8 +2,21 @@
 History
 =======
 
-1.1.0
------
+1.1.2 (Unreleased)
+------------------
+
+
+1.1.1 (2016-10-04)
+------------------
+
+* BODY-54: Improve resource events to enable sqs message queue and workflow events. (rudaporto)
+* Move validate_id function to validator module and fix tests. (rudaporto)
+* BODY-58: Avoid transaction rollback when connection to internal user service fails. (rudaporto)
+* When id field is in the collection_post body, check if alredy exists a registry with same id. (rudaporto)
+
+
+1.1.0 (2016-09-27)
+------------------
 
 * Create groupfinder callback and add it Authentication policy (JWT). (rudaporto)
 * Create new class to represent the current Authenticated user object in request. (rudaporto)
@@ -29,9 +42,10 @@ History
 * Fix workflow transition to not crash with empty message on body. (rudaporto)
 
 
-0.1.0 (Unreleased)
-------------------
+0.1.0 (internal release)
+------------------------
 
 * /__lbheartbeat__ endpoint to be used with load balancers. (ericof)
 * Error views for 403, 404 and 50x status codes. (ericof)
 * Initial implementation. (ericof)
+